+++ android/src/main/java/com/freeraspreactnative/FreeraspReactNativeModule.kt
@@ -4,41 +4,75 @@ import com.aheaditec.talsec_security.security.api.Talsec
 import com.aheaditec.talsec_security.security.api.TalsecConfig
 import com.aheaditec.talsec_security.security.api.ThreatListener
 import com.facebook.react.bridge.Arguments
+import com.facebook.react.bridge.Promise
 import com.facebook.react.bridge.ReactApplicationContext
 import com.facebook.react.bridge.ReactContextBaseJavaModule
 import com.facebook.react.bridge.ReactMethod
-import com.facebook.react.bridge.WritableMap
 import com.facebook.react.bridge.ReadableMap
+import com.facebook.react.bridge.WritableArray
 import com.facebook.react.modules.core.DeviceEventManagerModule
 
 class FreeraspReactNativeModule(val reactContext: ReactApplicationContext) :
-  ReactContextBaseJavaModule(reactContext), ThreatListener.ThreatDetected,
-  FreeraspDeviceStateListener.DeviceStateListener {
+  ReactContextBaseJavaModule(reactContext) {
 
-  private val listener = ThreatListener(this, FreeraspDeviceStateListener)
+  private val listener = ThreatListener(FreeraspThreatHandler, FreeraspThreatHandler)
 
   override fun getName(): String {
     return NAME
   }
 
+  init {
+    appReactContext = reactContext
+  }
+
   @ReactMethod
   fun talsecStart(
-    options: ReadableMap
+    options: ReadableMap,
+    promise: Promise
   ) {
 
     try {
       val config = parseTalsecConfig(options)
-      FreeraspDeviceStateListener.listener = this
+      FreeraspThreatHandler.listener = ThreatListener
       listener.registerListener(reactContext)
       Talsec.start(reactContext, config)
-      sendOngoingPluginResult("started", null)
+      promise.resolve("freeRASP started")
 
     } catch (e: Exception) {
       val params = Arguments.createMap().apply {
         putString("message", e.message)
       }
-      sendOngoingPluginResult("initializationError", params)
+      promise.reject("initializationError", params)
     }
+  }
+
+  /**
+   * Method to get the random identifiers of callbacks
+   */
+  @ReactMethod
+  fun getThreatIdentifiers(promise: Promise) {
+    promise.resolve(Threat.getThreatValues())
+  }
+
+  /**
+   * Method to setup the message passing between native and React Native
+   * @return list of [THREAT_CHANNEL_NAME, THREAT_CHANNEL_KEY]
+   */
+  @ReactMethod
+  fun getThreatChannelData(promise: Promise) {
+    val channelData: WritableArray = Arguments.createArray()
+    channelData.pushString(THREAT_CHANNEL_NAME)
+    channelData.pushString(THREAT_CHANNEL_KEY)
+    promise.resolve(channelData)
+  }
+
+  /**
+   * We never send an invalid callback over our channel.
+   * Therefore, if this happens, we want to kill the app.
+   */
+  @ReactMethod
+  fun onInvalidCallback() {
+    android.os.Process.killProcess(android.os.Process.myPid())
   }
 
   @ReactMethod
@@ -49,48 +83,6 @@ class FreeraspReactNativeModule(val reactContext: ReactApplicationContext) :
   @ReactMethod
   fun removeListeners(@Suppress("UNUSED_PARAMETER") count: Int) {
     // Remove upstream listeners, stop unnecessary background tasks
-  }
-
-  override fun onRootDetected() {
-    sendOngoingPluginResult("privilegedAccess", null)
-  }
-
-  override fun onDebuggerDetected() {
-    sendOngoingPluginResult("debug", null)
-  }
-
-  override fun onEmulatorDetected() {
-    sendOngoingPluginResult("simulator", null)
-  }
-
-  override fun onTamperDetected() {
-    sendOngoingPluginResult("appIntegrity", null)
-  }
-
-  override fun onUntrustedInstallationSourceDetected() {
-    sendOngoingPluginResult("unofficialStore", null)
-  }
-
-  override fun onHookDetected() {
-    sendOngoingPluginResult("hooks", null)
-  }
-
-  override fun onDeviceBindingDetected() {
-    sendOngoingPluginResult("deviceBinding", null)
-  }
-
-  override fun onObfuscationIssuesDetected() {
-    sendOngoingPluginResult("obfuscationIssues", null)
-  }
-
-  override fun deviceStateChangeDetected(threatType: String) {
-    sendOngoingPluginResult(threatType, null)
-  }
-
-  private fun sendOngoingPluginResult(eventName: String, params: WritableMap?) {
-    reactContext
-      .getJSModule(DeviceEventManagerModule.RCTDeviceEventEmitter::class.java)
-      .emit(eventName, params)
   }
 
   private fun parseTalsecConfig(config: ReadableMap): TalsecConfig {
@@ -125,5 +117,23 @@ class FreeraspReactNativeModule(val reactContext: ReactApplicationContext) :
 
   companion object {
     const val NAME = "FreeraspReactNative"
+    val THREAT_CHANNEL_NAME = (10000..999999999).random()
+      .toString() // name of the channel over which threat callbacks are sent
+    val THREAT_CHANNEL_KEY = (10000..999999999).random()
+      .toString() // key of the argument map under which threats are expected
+    private lateinit var appReactContext: ReactApplicationContext
+    private fun notifyListeners(threat: Threat) {
+      val params = Arguments.createMap()
+      params.putInt(THREAT_CHANNEL_KEY, threat.value)
+      appReactContext
+        .getJSModule(DeviceEventManagerModule.RCTDeviceEventEmitter::class.java)
+        .emit(THREAT_CHANNEL_NAME, params)
+    }
+  }
+
+  internal object ThreatListener : FreeraspThreatHandler.TalsecReactNative {
+    override fun threatDetected(threatType: Threat) {
+      notifyListeners(threatType)
+    }
   }
 }