+++ ios/FreeraspReactNative.m
@@ -4,7 +4,14 @@
 
 @interface RCT_EXTERN_MODULE(FreeraspReactNative, RCTEventEmitter)
 
-RCT_EXTERN_METHOD(talsecStart:(NSDictionary*)options)
+RCT_EXTERN_METHOD(talsecStart:(NSDictionary*)options
+                  withResolver:(RCTPromiseResolveBlock)resolve
+                  withRejecter:(RCTPromiseRejectBlock)reject)
+RCT_EXTERN_METHOD(getThreatChannelData:(RCTPromiseResolveBlock)resolve
+                  withRejecter:(RCTPromiseRejectBlock)reject)
+RCT_EXTERN_METHOD(getThreatIdentifiers:(RCTPromiseResolveBlock)resolve
+                  withRejecter:(RCTPromiseRejectBlock)reject)
+RCT_EXTERN_METHOD(onInvalidCallback)
 
 + (BOOL)requiresMainQueueSetup
 {